+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 
 - Set substance units on Nodes.
 
+- Set pumps on start node and add line geometry in geojson export.
+
 
 2.2.6 (2024-03-13)
 ------------------